fix: remove OpenSSL dependency, fix macOS cross-compilation
- Fix doctor.rs f_bsize/f_bavail type mismatch on macOS (cast to u64)
- Change shannon-tools reqwest to workspace inheritance, eliminating
  the last source of default-tls/native-tls/openssl-sys
- Regenerate Cargo.lock without OpenSSL dependencies

+++ crates/shannon-core/src/doctor.rs
@@ -806,7 +806,7 @@ fn get_disk_space(path: &std::path::Path) -> Option<u64> {
         // by the FFI call. We use `assume_init_read()` to move the initialized value out.
         let stat = unsafe { stat.assume_init_read() };
         // Available space = block size * available blocks
-        Some(stat.f_bsize * stat.f_bavail)
+        Some(stat.f_bsize as u64 * stat.f_bavail as u64)
     } else {
         // statvfs failed; the stat struct may be uninitialized, so we must not read it.
         // Returning None is the correct error handling path.
